@@ -6,27 +6,28 @@
   "scripts": {
     "start": "webpack-dev-server -d",
     "build": "webpack --progress -p",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "babel-node test"
   },
   "keywords": [],
   "author": "<git@tkel.ly>",
   "license": "AGPL-3.0+",
   "devDependencies": {
-    "babel-core": "^6.22.1",
-    "babel-loader": "^6.2.10",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
     "babel-polyfill": "^6.22.0",
-    "babel-preset-env": "^1.1.8",
-    "babel-preset-react": "^6.22.0",
-    "babel-preset-stage-2": "^6.22.0",
-    "css-loader": "^0.26.1",
-    "file-loader": "^0.10.0",
-    "postcss-cssnext": "^2.9.0",
-    "postcss-loader": "^1.2.2",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "css-loader": "^0.28.0",
+    "file-loader": "^0.11.1",
+    "postcss-cssnext": "^2.10.0",
+    "postcss-loader": "^1.3.3",
     "react-hot-loader": "^3.0.0-beta.6",
-    "style-loader": "^0.13.1",
+    "style-loader": "^0.16.1",
     "url-loader": "^0.5.7",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.4.1"
+    "webpack": "^2.4.1",
+    "webpack-dev-server": "^2.4.5"
   },
   "babel": {
     "presets": [
@@ -53,9 +54,10 @@
     ]
   },
   "dependencies": {
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-google-maps": "^6.0.1",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-google-maps": "^6.3.0",
     "react-icons": "^2.2.3",
     "xml2js": "^0.4.17"
   }
